0.5.0 pvm updates (#232)
* update instructions

* fix decode imm

* fix decode imm

* updates

* disable pvm test

+++ PolkaVM/Sources/PolkaVM/Registers.swift
@@ -20,23 +20,23 @@ public struct Registers: Equatable {
         }
     }
 
-    public var reg1: UInt32 = 0
-    public var reg2: UInt32 = 0
-    public var reg3: UInt32 = 0
-    public var reg4: UInt32 = 0
-    public var reg5: UInt32 = 0
-    public var reg6: UInt32 = 0
-    public var reg7: UInt32 = 0
-    public var reg8: UInt32 = 0
-    public var reg9: UInt32 = 0
-    public var reg10: UInt32 = 0
-    public var reg11: UInt32 = 0
-    public var reg12: UInt32 = 0
-    public var reg13: UInt32 = 0
+    public var reg1: UInt64 = 0
+    public var reg2: UInt64 = 0
+    public var reg3: UInt64 = 0
+    public var reg4: UInt64 = 0
+    public var reg5: UInt64 = 0
+    public var reg6: UInt64 = 0
+    public var reg7: UInt64 = 0
+    public var reg8: UInt64 = 0
+    public var reg9: UInt64 = 0
+    public var reg10: UInt64 = 0
+    public var reg11: UInt64 = 0
+    public var reg12: UInt64 = 0
+    public var reg13: UInt64 = 0
 
     public init() {}
 
-    public init(_ values: [UInt32]) {
+    public init(_ values: [UInt64]) {
         assert(values.count == 13)
         reg1 = values[0]
         reg2 = values[1]
@@ -55,13 +55,13 @@ public struct Registers: Equatable {
 
     /// standard program init registers
     public init(config: DefaultPvmConfig, argumentData: Data?) {
-        self[Index(raw: 0)] = UInt32(config.pvmProgramInitRegister1Value)
-        self[Index(raw: 1)] = UInt32(config.pvmProgramInitStackBaseAddress)
-        self[Index(raw: 7)] = UInt32(config.pvmProgramInitInputStartAddress)
-        self[Index(raw: 8)] = UInt32(argumentData?.count ?? 0)
+        self[Index(raw: 0)] = UInt64(config.pvmProgramInitRegister1Value)
+        self[Index(raw: 1)] = UInt64(config.pvmProgramInitStackBaseAddress)
+        self[Index(raw: 7)] = UInt64(config.pvmProgramInitInputStartAddress)
+        self[Index(raw: 8)] = UInt64(argumentData?.count ?? 0)
     }
 
-    public subscript(index: Index) -> UInt32 {
+    public subscript(index: Index) -> UInt64 {
         get {
             switch index.value {
             case 0:
